+++ recipe/key/laravel.php
@@ -7,4 +7,4 @@ require_once 'recipe/laravel.php';
 set('key_platform', 'LARAVEL');
 
 require_once __DIR__ . '/../key.php';
-require_once __DIR__ . '/laravel/build.php';
+require_once __DIR__ . '/../helpers/build.php';